Return pandas Series from scaled() and DataFrame from reduced()
DWChannel.reduced() returns a time series DataFrame with columns
ave, min, max, and rms. The advantage is that this is immediately
useful for postprocessing.

Similarly, DWChannel.scaled() now returns a Series instead of two
arrays for time and data.

As before, DWChannel.series() returns a series: scaled, if available,
or the ave data from reduced.